Fixed Compilation Bug in LString

+++ Experio/Source/Runtime/Containers/LString.h
@@ -57,7 +57,7 @@ std::string LString::NumberWithCommas(Number n)
 		return result;
 	}
 
-	n = LMath::Abs(n);
+	n = LMath::Abs((float)n);
 	std::string numberString = std::to_string(n);
 	for (int i = numberString.size() - 1; i >= 0; i--)
 	{
